requirements.txt
bleak>=0.21.0
pyserial>=3.5  # only needed for the serial_fallback.py
